DASH-264 This commit attempts to solve two related problems.
First, the commit attempts to execute database inserts of news-items, news-links, calendar-items and calendar-links in transactions such that the inserts will be committed before further database queries occur that might depend on those inserts.  This is part of the effort to eliminate contention seen in recent load tests of dashboard.  Further tuning may be possible by setting the isolation and/or propagation properties on the the transaction manager.  If anybody has advice about better settings than the defaults, we should try them. 

Second, the commit attempts to use batch-updates to insert collections of news-links and calendar-links.  We believe this will improve the performance of the dashboard's event-processing thread.  


git-svn-id: https://source.sakaiproject.org/svn/dashboard/trunk@313463 66ffb92e-73f9-0310-93c1-f5514f145a0a